+++ io_mesh_w3d/common/structs/mesh_structs/triangle.py
@@ -2,8 +2,8 @@
 # Written by Stephan Vedder and Michael Schnabel
 
 from mathutils import Vector
-from io_mesh_w3d.w3d.io_binary import *
-from io_mesh_w3d.w3x.io_xml import *
+from ....w3d.io_binary import *
+from ....w3x.io_xml import *
 
 
 surface_types = [